\Add SMULL instruction support and fix 16-bit parser types
- Introduce `smull_32` opcode in `thumb_2_opcodes.d`
- Add `parse_smull_32` and `execute_smull_32` functions
- Update `decode_instr` and `execute_instr` to handle `SMULL`
- Add `smull_32` to `opcode_strings`
- Include new test file `thumb_2_long_mult_acc_div_32.d` in `dub.json`
- Fix parsing functions for 16-bit instructions to use `ushort` instead of `short`
- Minor memory section update to handle UART writes at `0x40004404`

+++ src/thumb_2_load_store_single_data_item_16.d
@@ -18,7 +18,7 @@ enum field_tuples_strh_reg = [Tuple!(opcode, string[])(opcode.strh_reg, ["rt","r
 	STRH <Rt>,[<Rn>,<Rm>]
 	[15:9] 0101001, [8:6] Rm, [5:3] Rn, [2:0] Rt
 */
-instr_16 parse_strh_reg(uint instr) {
+instr_16 parse_strh_reg(ushort instr) {
 	instr_16 res;
 	res.op   = opcode.strh_reg;
 	ubyte rt = cast(ubyte)( instr       & 0x7);
